+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.1.2] - 2018-06-16
+### Added
+- Added Engine class for iterating order processes
+- Added a Process wrapper for functions being passed to the Engine
+
 ## [0.1.1] - 2018-06-12
 ### Added
 - Added cancelOrders() method on broker to cancel orders whose limit price differs from the best price on the orderbook